separate process of compiling jades outside of dev config
