+++ server/server.js
@@ -2,10 +2,12 @@
 
 import path from 'path'
 import bodyParser from 'body-parser'
+import express from 'express'
 import cookieParser from 'cookie-parser'
 import cookieSession from 'cookie-session'
+
 import getSettings from './getSettings'
-import express from 'express'
+import indexTemplate from './renderIndexTemplate'
 import { getPassport, addAuth } from './auth'
 
 import webpack from 'webpack'
@@ -16,6 +18,7 @@ import config from '../config/webpack/dev.js'
 const settings = getSettings()
 const app = express()
 const passport = getPassport(settings)
+const router = express.Router()
 
 if(process.env.NODE_ENV !== 'development') {
     app.use('/', express.static(path.resolve(__dirname, '..', 'dist')));
@@ -27,6 +30,12 @@ if(process.env.NODE_ENV !== 'development') {
     app.use(webpackMiddleware(compiler));
     app.use(webpackHotMiddleware(compiler));
 }
+
+router.get('/', (req, res, next) => {
+    res.statusCode = 200;
+    res.setHeader('Content-Type', 'text/html')   
+    res.end(indexTemplate)
+} )
 
 app.use(cookieParser());
 app.use(bodyParser.urlencoded({extended: true}));
